+++ frontend/appflowy_flutter/lib/mobile/presentation/database/field/bottom_sheet_create_field.dart
@@ -1,13 +1,14 @@
 import 'package:appflowy/mobile/presentation/bottom_sheet/bottom_sheet.dart';
-import 'package:appflowy/mobile/presentation/database/card/card_detail/mobile_create_field_screen.dart';
-import 'package:appflowy/mobile/presentation/database/card/card_detail/mobile_edit_field_screen.dart';
-import 'package:appflowy/mobile/presentation/database/card/card_detail/widgets/_field_options.dart';
-import 'package:appflowy/mobile/presentation/database/card/card_detail/widgets/_field_options_eidtor.dart';
+import 'package:appflowy/mobile/presentation/database/field/mobile_field_type_grid.dart';
+import 'package:appflowy/mobile/presentation/database/field/mobile_field_type_option_editor.dart';
 import 'package:appflowy/mobile/presentation/database/field/quick_edit_field_bottom_sheet.dart';
 import 'package:appflowy/plugins/database_view/application/field/field_info.dart';
 import 'package:appflowy/plugins/database_view/application/field/field_service.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
+
+import 'mobile_create_field_screen.dart';
+import 'mobile_edit_field_screen.dart';
 
 void showCreateFieldBottomSheet(BuildContext context, String viewId) {
   showMobileBottomSheet(
